bouffalolab_release_bl_iot_sdk_1.6.40-2000-gb17d51bc3

+++ utils/src/utils_string.c
@@ -59,7 +59,7 @@ void get_bytearray_from_string(char** params, uint8_t *result,int array_size)
 
 void get_uint8_from_string(char** params, uint8_t *result)
 {
-	uint32_t p = 0;
+	uint64_t p = 0;
 	int state=0;
 	
 	state = params_filter(params,&p);
@@ -71,7 +71,7 @@ void get_uint8_from_string(char** params, uint8_t *result)
 
 void get_uint16_from_string(char** params, uint16_t *result)
 {
-	uint32_t p = 0;
+	uint64_t p = 0;
 	int state=0;
 	
 	state = params_filter(params,&p);
@@ -83,7 +83,7 @@ void get_uint16_from_string(char** params, uint16_t *result)
 
 void get_uint32_from_string(char** params, uint32_t *result)
 {
-    uint32_t p = 0;
+    uint64_t p = 0;
 	int state=0;
 	
 	state = params_filter(params,&p);
